Update day 2

+++ Day2/main.cpp
@@ -4,16 +4,16 @@
 #include <vector>
 
 
-int read_depth() {
+size_t read_depth() {
     std::string input;
     std::ifstream input_file("input.txt");
 
     std::vector<int> measurements;
-    int count = 0;
+    size_t count = 0;
     while (std::getline(input_file, input)) {
         measurements.push_back(std::stoi(input));
 
-        int i = measurements.size() - 1;
+        size_t i = measurements.size() - 1;
         if(i > 2){
             int prev_sum = measurements[i - 1] + measurements[i - 2] + measurements[i - 3];
             int current_sum = measurements[i] + measurements[i - 1] + measurements[i - 2];
@@ -28,7 +28,7 @@ int read_depth() {
 }
 
 int main() {
-    int count = read_depth();
+    size_t count = read_depth();
     std::cout << count << std::endl;
     return 0;
 }
